android-daemon: Remove network-state listener from AndroidManifest
The listener for network-states is removed to prevent unnecessary launches of 
the service while these are turned off. Instead the service instantiates a BroadcastReceiver
as soon as the daemon is in online state.
